feat: added ChangeBookQuantityInCartDto. Moved CartController from cart to open
