+++ src/main/java/teamchallenge/server/dto/ChangeBookQuantityInCartDto.java
@@ -1,0 +1,10 @@
+package teamchallenge.server.dto;
+
+import lombok.Data;
+
+@Data
+public class ChangeBookQuantityInCartDto {
+    String token;
+    Long productId;
+    Long quantity;
+}
